HOTT-1768: Add namespace to facet configuration model
