+++ spec/models/beta/search/search_facet_classifier_configuration_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe SearchFacetClassifierConfiguration do
+RSpec.describe Beta::Search::SearchFacetClassifierConfiguration do
   describe '.each_classification' do
     context 'when passed a block' do
       it { expect { |block| described_class.each_classification(&block) }.to yield_control.exactly(69).times }
@@ -11,5 +11,11 @@ RSpec.describe SearchFacetClassifierConfiguration do
 
       it { is_expected.to be_nil }
     end
+  end
+
+  describe '.heading_facet_mappings' do
+    subject(:heading_facet_mappings) { described_class.heading_facet_mappings.keys.count }
+
+    it { is_expected.to eq(1230) }
   end
 end
